% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartmix-package.R
\docType{package}
\name{bartmix-package}
\alias{bartmix}
\alias{bartmix-package}
\title{bartmix: Environmental Mixture Analysis with BART}
\description{
Functions for environmental mixture analyses with Bayesian additive regression trees (BART). The included functions do posterior analyses that may be of interested in environmental mixture analyses. Many of the functions mimic what is typically done with Bayesian kernel machine regression. The package is designed to use output from BART models fit with the BART package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnderWilson/bartmix}
  \item Report bugs at \url{https://github.com/AnderWilson/bartmix/issues}
}

}
\author{
\strong{Maintainer}: Ander Wilson \email{ander.wilson@colostate.edu} (\href{https://orcid.org/0000-0003-4774-3883}{ORCID})

}
\keyword{internal}
