% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialdependence1.R
\name{partialdependence1}
\alias{partialdependence1}
\title{1-way partial dependence}
\usage{
partialdependence1(object, data, exposures, L = 50)
}
\arguments{
\item{object}{A fitted object of class wbart from the BART R package.}

\item{data}{The data used to fit the BART model.}

\item{exposures}{Names of exposures to calculate the partial dependence function for. This can be the name of a single exposure or a vector of names.}

\item{L}{The number of points to estimate the exposure-response function of var at.}
}
\value{
A data.frame with the estimated partial dependence function(s).
}
\description{
1-way partial dependence
}
