% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialdependence1.R
\name{partialdependence1_helper}
\alias{partialdependence1_helper}
\title{1-way partial dependence for a single exposure}
\usage{
partialdependence1_helper(object, data, varname, L = 50)
}
\arguments{
\item{object}{A fitted object of class wbart from the BART R package.}

\item{data}{The data used to fit the BART model.}

\item{varname}{Name of exposure  to calculate the partial dependence function for.}

\item{L}{The number of points to estimate the exposure-response function of var at.}
}
\value{
A data.frame with the estimated exposure-response function.
}
\description{
1-way partial dependence for a single exposure
}
