% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialdependence2.R
\name{partialdependence2}
\alias{partialdependence2}
\title{2-way partial dependence}
\usage{
partialdependence2(object, data, var, var2, qtls = c(0.25, 0.5, 0.75), L = 50)
}
\arguments{
\item{object}{A fitted object of class wbart from the BART R package.}

\item{data}{The data used to fit the BART model.}

\item{var}{The name of the exposure of interest in data. The exposure-response function for this variable will be produced}

\item{var2}{The name of a second exposure variable. The exposure-response function for var will be produced at specified quantiles of var2.}

\item{qtls}{The quantiles of var2 to be used.}

\item{L}{The number of points to estimate the exposure-response function of var at.}
}
\value{
A data.frame with the exposure-response functions
}
\description{
2-way partial dependence
}
