% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalmixtureeffect.R
\name{totalmixtureeffect}
\alias{totalmixtureeffect}
\title{Total mixture effect}
\usage{
totalmixtureeffect(
  object,
  data,
  exposures = colnames(data),
  qtls = seq(0.2, 0.8, 0.05)
)
}
\arguments{
\item{object}{A fitted object of class wbart from the BART R package.}

\item{data}{The data used to fit the BART model.}

\item{exposures}{Vector with the variable names of exposures in the data.}

\item{qtls}{A vector of quantiles to estimate the total mixture effect at.}
}
\value{
A data.frame with the total mixture effect.
}
\description{
Total mixture effect
}
